finished videoplayer

+++ controllers/videoController.js
@@ -1,27 +1,27 @@
 import routes from "../routes";
 import Video from "../models/video";
 
-export const home = async(req, res) => {
+export const home = async (req, res) => {
     try {
-        const videos = await Video.find({}).sort({ _id: -1});
+        const videos = await Video.find({}).sort({ _id: -1 });
         res.render("home", { pageTitle: "Home", videos });
-    } catch(error){
+    } catch (error) {
         console.log(error);
         res.render("home", { pageTitle: "Home", videos: [] });
     }
 
 };
 
-export const search = async(req, res) => {
+export const search = async (req, res) => {
     const {
         query: { term: searchingBy }
     } = req;
     let videos = [];
-    try{
-        videos= await Video.find({
-            title: {$regex: searchingBy, $options: "i"}
+    try {
+        videos = await Video.find({
+            title: { $regex: searchingBy, $options: "i" }
         });
-    } catch(error) {
+    } catch (error) {
         console.log(error)
     }
     res.render("search", { pageTitle: "Search", searchingBy, videos });
@@ -30,8 +30,8 @@ export const search = async(req, res) => {
 export const getUpload = (req, res) =>
     res.render("upload", { pageTitle: "Upload" });
 
-export const postUpload = async(req, res) => {
-    const { 
+export const postUpload = async (req, res) => {
+    const {
         body: { title, description },
         file: { path }
     } = req;
@@ -40,61 +40,72 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
-    })
+        description,
+        creator: req.user.id
+    });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     console.log(newVideo);
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
 
-export const videoDetail = async(req, res) => {
+export const videoDetail = async (req, res) => {
     const {
-        params: {id}
+        params: { id }
     } = req;
     try {
-        const video = await Video.findById(id);
-        res.render("videoDetail", { pageTitle: Video.title, video });
+        const video = await Video.findById(id).populate("creator");
         console.log(video);
+        res.render("videoDetail", { pageTitle: Video.title, video });
     } catch (error) {
         res.redirect(routes.home);
     }
 };
-export const getEditVideo = async(req, res) => {
+export const getEditVideo = async (req, res) => {
     const {
         params: { id }
     } = req;
     try {
         const video = await Video.findById(id);
         console.log(video);
-        res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
-    } catch(error) {
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
+    } catch (error) {
         res.redirect(routes.home);
     }
     res.render("editVideo", { pageTitle: "Edit Video" });
 }
 
-export const postEditVideo = async(req, res) => {
+export const postEditVideo = async (req, res) => {
     const {
-        params:{id},
-        body: {title, description}
+        params: { id },
+        body: { title, description }
     } = req;
     try {
-        await Video.findOneAndUpdate({ _id: id }, {title, description});
+        await Video.findOneAndUpdate({ _id: id }, { title, description });
         res.redirect(routes.videoDetail(id));
-    } catch(error) {
+    } catch (error) {
         res.redirect(routes.home);
     }
 };
 
-export const deleteVideo = async(req, res) => {
+export const deleteVideo = async (req, res) => {
     const {
-        params: {id}
+        params: { id }
     } = req;
 
-    try{
-        await Video.findOneAndRemove({_id: id});
-        // res.render("deleteVideo", { pageTitle: "Delete Video" });
-    } catch(error) {
+    try {
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
+    } catch (error) {
 
     }
     res.redirect(routes.home);
